refactor: use unified getNumberConfig function for PARALLEL_LIMIT per code review
* Move PARALLEL_LIMIT from workflow constants to config/index.ts using getNumberConfig helper
* Enhance getNumberConfig with NaN validation and positive number checks
* Update iteration panel to import PARALLEL_LIMIT from @/config instead of constants
* Maintain all error handling and edge cases in the unified configuration system
* Update tests to reflect new import path while preserving all test coverage

Addresses code review feedback: Use existing getNumberConfig infrastructure instead
of duplicating environment variable parsing logic in constants.ts
