+++ web/app/components/workflow/nodes/iteration/panel.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 import VarReferencePicker from '../_base/components/variable/var-reference-picker'
 import Split from '../_base/components/split'
-import { MIN_ITERATION_PARALLEL_NUM, PARALLEL_LIMIT } from '../../constants'
+import { MIN_ITERATION_PARALLEL_NUM } from '../../constants'
 import type { IterationNodeType } from './types'
 import useConfig from './use-config'
 import { ErrorHandleMode, type NodePanelProps } from '@/app/components/workflow/types'
@@ -12,6 +12,7 @@ import Switch from '@/app/components/base/switch'
 import Select from '@/app/components/base/select'
 import Slider from '@/app/components/base/slider'
 import Input from '@/app/components/base/input'
+import { PARALLEL_LIMIT } from '@/config'
 
 const i18nPrefix = 'workflow.nodes.iteration'
